### 2018-09-17 #958 @arfon
@jrosen48 - please take a look at these example `paper.md` and `paper.bib` files and format your paper in exactly the same way please: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
