### 2018-09-17 e35cf34
update paper.md and paper.bib
