### 2018-09-17 d332e31
delete extra paper docs
