### 2018-09-17 #958 @arfon
Please modify your `paper.md` file to have an `affiliations` section as per the example here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
