### 2018-09-17 #958 @jrosen48
I've updated the `paper.md` file to have the `affiliations` section in the right format. I'm sorry again about that; thank you for walking me through this step that I should have had correct when I submitted.
