### 2018-09-17 #958 @jrosen48
I'd like to suggest the following as potential reviewers (if they are available and willing):

- @strengejacke
- @jnese (if I understand correctly through a colleague, he uses [the commercial software] MPlus and can test functions the subset of functions in the package that require MPlus)
