### 2018-09-21 60c3748
update model print out
