+++ man/compare_solutions_mplus.Rd
@@ -58,7 +58,7 @@ Explore the BIC values of a range of Mplus models in terms of a) the structure o
 }
 \examples{
 \dontrun{
-compare_solutions_mplus(iris, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
+o <- compare_solutions_mplus(iris, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
 n_profiles_max = 3)
 }
 }
