+++ man/compare_solutions_mplus.Rd
@@ -10,7 +10,8 @@ compare_solutions_mplus(df, ..., n_profiles_min = 2,
   10), cluster_ID = NULL, m_iterations = 500, st_iterations = 20,
   convergence_criterion = 1e-06, remove_tmp_files = FALSE,
   save_models = FALSE, return_table = TRUE, n_processors = 1,
-  return_stats_df = TRUE, include_VLMR = TRUE, include_BLRT = FALSE)
+  return_stats_df = TRUE, include_VLMR = TRUE, include_BLRT = FALSE,
+  dir_name = NULL)
 }
 \arguments{
 \item{df}{data.frame with two or more columns with continuous variables}
@@ -46,6 +47,8 @@ compare_solutions_mplus(df, ..., n_profiles_min = 2,
 \item{include_VLMR}{whether to include the Vu-Lo-Mendell-Rubin likelihood-ratio test; defaults to TRUE}
 
 \item{include_BLRT}{whether to include the bootstrapped LRT; defaults to FALSE because of the time this takes to run}
+
+\item{dir_name}{character; name for directory .out files are saved to if save_models = TRUE}
 }
 \value{
 a list with a data.frame with the BIC values and a list with all of the model output; if save_models is the name of an rds file (i.e., "out.rds"), then the model output will be written with that filename and only the data.frame will be returned
