### 2018-09-22 e1ca922
update helpers
