### 2018-09-23 135e499
update message about how to specify models
