### 2018-09-25 #958 @jrosen48
Hi @leeper - if it's alright, hoping to check-in regarding the next steps. Thank you very much.
