### 2018-09-25 #978 @whedon
[REVIEW]: Easily Carry Out Latent Profile Analysis (LPA) Using Open-Source or Commercial Software
**Submitting author:** @jrosen48 (<a href="http://orcid.org/0000-0003-2170-0447">Joshua Rosenberg</a>)
**Repository:** <a href="https://github.com/jrosen48/tidyLPA" target ="_blank">https://github.com/jrosen48/tidyLPA</a>
**Version:** 0.2.1 (development version; most recent CRAN version is 0.2.0)

- [x] **Version:** Does the release version given match the GitHub release (0.2.1 (development version; most recent CRAN version is 0.2.0))?

- [x] **Version:** Does the release version given match the GitHub release (0.2.1 (development version; most recent CRAN version is 0.2.0))?
