### 2018-09-25 79f6e36
change AIC stat to fix bug
