+++ R/compare-solutions-mplus.R
@@ -150,7 +150,7 @@ compare_solutions_mplus <- function(df, ...,
             cluster_ID = cluster_ID_label,
             LL = m$summaries$LL,
             npar = m$summaries$Parameters,
-            AIC = m$summaries$LL,
+            AIC = m$summaries$AIC,
             BIC = m$summaries$BIC,
             SABIC = m$summaries$aBIC,
             CAIC = m$summaries$AICC,
@@ -174,7 +174,7 @@ compare_solutions_mplus <- function(df, ...,
             cluster_ID = cluster_ID_label,
             LL = m$summaries$LL,
             npar = m$summaries$Parameters,
-            AIC = m$summaries$LL,
+            AIC = m$summaries$AIC,
             BIC = m$summaries$BIC,
             SABIC = m$summaries$aBIC,
             CAIC = m$summaries$AICC,
