### 2018-09-26 jrosen48/tidyLPA#70 @strengejacke
[JOSS-Review] fix license.md
This issue is related to https://github.com/openjournals/joss-reviews/issues/978.

Just a minor thing, please change "your name goes here" in https://github.com/jrosen48/tidyLPA/blob/master/LICENSE.md.
