### 2018-09-26 jrosen48/tidyLPA#71 @strengejacke
[JOSS-Review] Add missing dois to paper
This one is related to https://github.com/openjournals/joss-reviews/issues/978

There appear no DOIs in the paper's references. I assume that at least the cited journal articles have a DOI, so please check and add DOIs where available.
