### 2018-09-26 jrosen48/tidyLPA#72 @strengejacke
[JOSS-review] clarify docs
This one is related to https://github.com/openjournals/joss-reviews/issues/978

I just installed the package and tried out some functions. For `estimate_profiles()`, the help-file says, as return value
> either a tibble or a ggplot2 plot of the BIC values for the explored models

But I don't know how to create a ggplot-object. Could you clarify this in your docs, or help me finding the relevant passages in the help?
