### 2018-09-26 jrosen48/tidyLPA#74 @strengejacke
[JOSS-review] statement of need in readme.md
This one is related to https://github.com/openjournals/joss-reviews/issues/978

While there's a clear statement of need in the paper, the _background_ section of the readme.md could elaborate a bit more on why tidyLPA is needed, also in the context of what already exists. It's just a minor issue, I would say adding a small sentence to the background should be OK.
