### 2018-09-26 905899d
improve save_models functionality to create a directory name based on the data
