### 2018-09-26 c6eddb7
change how save models file names are constructed
