### 2018-09-28 jrosen48/tidyLPA#70 @jrosen48
Thank you @strengejacke for this and your other very helpful feedback. I'm looking forward to addressing.
