### 2018-09-28 a5c736d
Update LICENSE.md
add names to license in response to #70
