### 2018-09-28 jrosen48/tidyLPA#70 @jrosen48
Fixed this.
