### 2018-09-30 095acd9
add mplusavailable() function and functionality to check whether MPlus is installed when Mplus functions are called
