+++ paper.bib
@@ -8,7 +8,7 @@
     volume = {8},
     number = {1},
     pages = {205--233},
-    adsurl =
+    url =
       {https://journal.r-project.org/archive/2017/RJ-2017-008/RJ-2017-008.pdf},
 }
 
@@ -27,39 +27,39 @@
     year = {2018},
     pages = {1--18},
     doi = {10.1080/10705511.2017.1402334},
-    adsurl =
+    url =
       {http://www.tandfonline.com/doi/full/10.1080/10705511.2017.1402334},
 }
 
 @article{pastor_et_al_2007,
-  Title={A latent profile analysis of college students’ achievement goal orientation},
-  Author={Pastor, Dena A and Barron, Kenneth E and Miller, BJ and Davis, Susan L},
-  Journal={Contemporary Educational Psychology},
-  Volume={32},
-  Number={1},
-  Pages={8--47},
-  Year={2007},
-  Publisher={Elsevier}
+  title={A latent profile analysis of college students’ achievement goal orientation},
+  author={Pastor, Dena A and Barron, Kenneth E and Miller, BJ and Davis, Susan L},
+  journal={Contemporary Educational Psychology},
+  volume={32},
+  number={1},
+  pages={8--47},
+  year={2007},
+  publisher={Elsevier}
 }
 
 @article{harring_et_al_2016,
-  Title={Mixture modeling: Applications in educational psychology},
-  Author={Harring, Jeffrey R and Hodis, Flaviu A},
-  Journal={Educational Psychologist},
-  Volume={51},
-  Number={3-4},
-  Pages={354--367},
-  Year={2016},
-  Publisher={Taylor \& Francis}
+  title={Mixture modeling: Applications in educational psychology},
+  author={Harring, Jeffrey R and Hodis, Flaviu A},
+  journal={Educational Psychologist},
+  volume={51},
+  number={3-4},
+  pages={354--367},
+  year={2016},
+  publisher={Taylor \& Francis}
 }
 
 @article{wickham14,
-  Title={Tidy data},
-  Author={Wickham, Hadley and others},
-  Journal={Journal of Statistical Software},
-  Volume={59},
-  Number={10},
-  Pages={1--23},
-  Year={2014},
-  Publisher={Foundation for Open Access Statistics}
+  title={Tidy data},
+  author={Wickham, Hadley and others},
+  journal={Journal of Statistical Software},
+  volume={59},
+  number={10},
+  pages={1--23},
+  year={2014},
+  publisher={Foundation for Open Access Statistics}
 }
