+++ paper.md
@@ -1,106 +1,37 @@
----
-title: 'tidyLPA: An R Package to Easily Carry Out Latent Profile Analysis (LPA) Using Open-Source or Commercial Software'
-tags:
-  - R
-  - mixture modeling
-  - Latent Profile Analysis
-  - MPlus
-authors:
-  - name: Joshua M. Rosenberg
-    orcid: 0000-0003-2170-0447
-    affiliation: 1
-  - name: Patrick N. Beymer
-    orcid: 0000-0002-3569-2305
-    affiliation: 2
-  - name: Daniel J. Anderson
-    orcid: 0000-0003-4699-4680
-    affiliation: 3
-  - name: Jennifer A. Schmidt
-    orcid: 0000-0003-2853-9801
-    affiliation: 2
-affiliations:
- - name: University of Tennessee, Knoxville
-   index: 1
- - name: Michigan State University
-   index: 2
- - name: University of Oregon
-   index: 3
-date: 16 September 2018
-bibliography: paper.bib
----
+tidyLPA
+================
+14 August 2018
 
-# Summary 
+tidyLPA: Easily carry out latent profile analysis
+=================================================
 
-Researchers are often interested in identifying homogeneous subgroups within
-heterogeneous samples on the basis of a set of measures, such as profiles of
-individuals' motivation (i.e., their values, competence beliefs, and achievement
-goals). Latent Profile Analysis (LPA) is a statistical method for identifying
-such groups, or *latent profiles*, and is a special case of the general
-mixture model where all measured variables are continuous [@harring_et_al_2016;
-@pastor_et_al_2007]. The **tidyLPA** package allows users to specify different
-models that determine whether and how different parameters (i.e., means,
-variances, and covariances) are estimated, and to specify and compare different
-solutions based on the number of profiles extracted.
+Researchers are often interested in identifying homogeneous subgroups within heterogeneous samples on the basis of a set of measures, such as dimensions of youths' engagement in educational settings (i.e., cognitively, behaviorally, and affectively). Latent Profile Analysis (LPA) is a statistical method for identifying such groups, or *latent profiles*, and is a special case of the general mixture model where all measured variables are continuous (Harring and Hodis 2016; Pastor et al. 2007). The **tidyLPA** package allows users to specify different models that determine whether and how different parameters (i.e., means, variances, and covariances) are estimated, and to specify and compare different solutions based on the number of profiles extracted.
 
-The aim of the **tidyLPA** package is to provide a simple interface for
-conducting and evaluating LPA models. Given that LPA is only one type of mixture
-model, we do not expect it to replace the more general functionality of other
-tools that allow for the estimation of wider range of models. Nevertheless, this
-package provides convenient methods for conducting LPA using both open-source
-and commercial software, while aligning with a widely used coding framework
-(i.e., *tidy* data, described more below). In doing so, **tidyLPA** allows
-researchers with and without access to proprietary tools, such as MPlus, to
-conduct LPA.
+The aim of the **tidyLPA** package is to provide a simple interface for conducting and evaluating LPA models. Given that LPA is only one type of mixture model, we do not expect it to replace the more general functionality of other tools that allow for the estimation of wider range of models. Nevertheless, this package provides convenient methods for conducting LPA using both open-source and commercial software, while aligning with a widely used coding framework (i.e., *tidy* data, described more below). In doing so, **tidyLPA** allows researchers with and without access to proprietary tools, such as MPlus, to conduct LPA.
 
-# A *tidy* user-interface
+A *tidy* user-interface
+-----------------------
 
-The input for *tidyLPA* assumes a tidy data structure [see @wickham14], and all
-output are returned in a tidy from, which aligns with the broad array of tools
[…]

[…]
