+++ NAMESPACE
@@ -6,6 +6,7 @@ export(compare_solutions_mplus)
 export(estimate_profiles)
 export(estimate_profiles_mplus)
 export(extract_LL_mplus)
+export(mplusAvailable)
 export(plot_profiles)
 export(plot_profiles_mplus)
 import(tibble)
