+++ README.md
@@ -1,6 +1,6 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-[![Build Status](https://travis-ci.org/jrosen48/tidyLPA.svg?branch=master)](https://travis-ci.org/jrosen48/tidyLPA) [![CRAN status](https://www.r-pkg.org/badges/version/tidyLPA)](https://cran.r-project.org/package=tidyLPA) [![](https://cranlogs.r-pkg.org/badges/tidyLPA)](https://cran.r-project.org/package=tidyLPA) [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
+[![Build Status](https://travis-ci.org/jrosen48/tidyLPA.svg?branch=master)](https://travis-ci.org/jrosen48/tidyLPA) [![CRAN status](https://www.r-pkg.org/badges/version/tidyLPA)](https://cran.r-project.org/package=tidyLPA) [![](https://cranlogs.r-pkg.org/badges/tidyLPA)](https://cran.r-project.org/package=tidyLPA) [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing) [![status](http://joss.theoj.org/papers/ccb887b4991abca9517847b016cdb33d/status.svg)](http://joss.theoj.org/papers/ccb887b4991abca9517847b016cdb33d)
 
 Background
 ----------
@@ -12,13 +12,13 @@ tidyLPA provides the functionality to carry out LPA in R. In particular, tidyLPA
 Installation
 ------------
 
-You can install tidyLPA (v. `0.2.0`) from CRAN with:
+You can install tidyLPA from CRAN with:
 
 ``` r
 install.packages("tidyLPA")
 ```
 
-You can also install the development version (v. `0.2.1`) of tidyLPA from GitHub with:
+You can also install the development version of tidyLPA from GitHub with:
 
 ``` r
 install.packages("devtools")
@@ -40,7 +40,7 @@ d <- pisaUSA15[1:100, ]
 estimate_profiles(d, 
                   broad_interest, enjoyment, self_efficacy, 
                   n_profiles = 3)
-#> Fit NA model with 3 profiles.
+#> Fit Equal variances and covariances fixed to 0 (model 1) model with 3 profiles.
 #> LogLik is 283.991
 #> BIC is 631.589
 #> Entropy is 0.914
