+++ README.Rmd
@@ -16,6 +16,7 @@ knitr::opts_chunk$set(
 [![CRAN status](https://www.r-pkg.org/badges/version/tidyLPA)](https://cran.r-project.org/package=tidyLPA)
 [![](https://cranlogs.r-pkg.org/badges/tidyLPA)](https://cran.r-project.org/package=tidyLPA)
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
+[![status](http://joss.theoj.org/papers/ccb887b4991abca9517847b016cdb33d/status.svg)](http://joss.theoj.org/papers/ccb887b4991abca9517847b016cdb33d)
 
 ## Background
 
