### 2018-10-02 de01c56
update paper for #71
