### 2018-10-02 jrosen48/tidyLPA#71 @jrosen48
Thank you for noticing this - I checked each reference and added DOIs. Every reference now has at least a URL or a DOI.
