### 2018-10-02 jrosen48/tidyLPA#74 @jrosen48
Thank you for this helpful issue! I added the following to the second paragraph of the **Background** section of the README:

> The package is designed and documented to be easy to use, especially for beginners to LPA, but with fine-grained options available for estimating models and evaluating specific output as part of more complex analyses.
