### 2018-10-02 jrosen48/tidyLPA#72 @jrosen48
Thank you for diving in and catching this. I've added the following line to the documentation to replace the line you highlighted:

> either a tibble with the data and profile/class assignment and the posterior probability for that profile; the output of this function can be passed to the plot_profiles() function to create a ggplot2 plot of the profiles. If the argument to_return = "mclust" is added to the function call) an mclust model object, which can be inspected or plotted with mclust functions.

One way you can plot the profiles is something like:

```{r}
estimate_profiles(iris,
                  Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
                  n_profiles = 3) %>%
    plot_profiles()
```

You could also (of course) save the output from `estimate_profiles()`, i.e.:

```{r}
m3 <- estimate_profiles(iris,
                  Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
                  n_profiles = 3)

plot_profiles(m3)
```

There are a few useful arguments to `plot_profiles()`, including the option to plot weighted (by the probability) means and to bootstrap the standard errors (h/t @cjvanlissa for this).
