### 2018-10-02 jrosen48/tidyLPA#72 @strengejacke
Looks good to me! Have you already committed the changes?
