### 2018-10-02 jrosen48/tidyLPA#72 @jrosen48
Just pushed--and made a minor change to the docs (changed in my comment above, too- I removed the word "either).
