### 2018-10-02 #978 @strengejacke
Ok, from my side all issues have been addressed adequately, so I have no further requests to the author(s).
