### 2018-10-02 c8c0aac
update compare_solutions_mplus()
