### 2018-10-02 #978 @jnese
Absolutely. I hope to have it completed in less time than that.
