### 2018-10-03 2a0a424
update labels for compare_solutions()
