### 2018-10-10 #978 @jrosen48
Thank you @strengejacke and @jnese for lending your time and expertise - the package is better and I appreciate the feedback you provided in terms of revisions to make after the paper is accepted, too! Thank you @arfon for your work on this, too - this was a helpful and illuminating peer-review experience and I appreciate your time and expertise as well.
