### 2018-10-10 #978 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
